% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{create_corpus}
\alias{create_corpus}
\title{Create a corpus}
\usage{
create_corpus(files, dbsource = "wos", format = "plaintext")
}
\arguments{
\item{files}{list of files}

\item{dbsource}{wos}

\item{format}{plaintext}
}
\description{
wrap around bibliometrix::convert2df
}
