% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{create_wordcloud}
\alias{create_wordcloud}
\title{Create a wordcloud from the co-occurring words}
\usage{
create_wordcloud(corpus = corpus, keywords = "plus")
}
\arguments{
\item{corpus}{corpus of literature}

\item{keywords}{either plus or author keywords used to make the wordcloud}
}
\description{
wrap around wordcloud2::wordcloud
}
